Make public API conform to standard
I am not sure about this change, however this is conforming to our
notifier API standard albeit not quite idiomatic Go.

Before this change the public API was the same as the internal API
meaning that depending on the arguments you have to call a different
function:

    Error
    ErrorWithExtras
    RequestErrorWithStackSkipWithExtras
    Message
    MessageWithExtras

et cetera. The first parameter was always the level, and the rest of the
name determined what arguments you pass. The alternative is what is
implemented here:

     Critical
     Error
     Warning
     Info
     Debug

which take `...interface{}` and `Log(level string, xs ...interface{})`
which they all delegate to with the specific level set. We then do
runtime type checking on the input parameters to call into the internal
API described above. This matches what our other notifiers do, but it
uses some gross Go doing the type casting and is arguably less clear to
the typical Go developer.