Retry temporary http errors and dump items to stderr
If an error occurs during transport such as a TLS timeout we attempt to
retry sending the item. We do this a fixed number of times. Once the
fixed number of attempts has exhausted we declare that a failure. We
also print the full payload to stderr or to the configured logger in the
event of a failure so at least the item shows up somewhere rather than
completely getting lost. Both the number of retries and whether or not
to print the item is configurable. The defaults are 3 retries and true
for printing.

We could arguably have more complex retry behaviour like adding some
backoff, but these are really only for very transient errors so adding
that complexity would be a next step.