Provide a Close mechanism to clean up the default transport goroutine
Fixes #48

By default when you import rollbar we create an async transport as part
of a default client so that the top level functions for reporting items
just work. However, that leads to a goroutine that dangles if you don't
actually use anything or don't want to use that standard client. This PR
adds `Close` to the top level to call Close on the standard client. We
also make `Close` close the underlying channel which is used for
communication to the goroutine used to process items. As a consequence,
once all items have been sent this goroutine will shutdown. So if you
never use it, Close will effectively just shut the goroutine down.